removed StakeExpenditureDialogForm
